+++ app/src/main/java/eu/kanade/presentation/track/TrackerSearch.kt
@@ -304,6 +304,15 @@ private fun SearchResultItem(
                             }
                         },
                     )
+                    if (trackSearch.authors.isNotEmpty() || trackSearch.artists.isNotEmpty()) {
+                        Text(
+                            text = (trackSearch.authors + trackSearch.artists).distinct().joinToString(),
+                            modifier = Modifier.secondaryItemAlpha(),
+                            maxLines = 1,
+                            overflow = TextOverflow.Ellipsis,
+                            style = MaterialTheme.typography.bodySmall,
+                        )
+                    }
                     if (type.isNotBlank()) {
                         SearchResultItemDetails(
                             title = stringResource(MR.strings.track_type),