+++ app/src/main/java/eu/kanade/tachiyomi/data/track/model/TrackSearch.kt
@@ -34,6 +34,10 @@ class TrackSearch : Track {
 
     override lateinit var tracking_url: String
 
+    var authors: List<String> = emptyList()
+
+    var artists: List<String> = emptyList()
+
     var cover_url: String = ""
 
     var summary: String = ""